Fix registrations issues
